Fix clippy failures after splitting Test Patterns UI.
Allow the settings panel helper's argument count and keep the scroll-phase unit test next to advance_scroll_phase.
